requirements.txt
streamlit>=1.28.0
yfinance>=0.2.31
pandas>=2.0.0
plotly>=5.18.0
firebase-admin>=6.2.0
requests>=2.31.0
numpy>=1.24.0
google-genai>=1.0.0
Pillow>=10.0.0
